Added bootloaders, Makefile allows making and booting kernel high or low, fixed a few bugs related to high kernel, simplified swapping. Fixed an assembler bug.

+++ trunk/cross-pcc/as/as26.c
@@ -153,7 +153,7 @@ void opline()
 	case TYPEOPASC:
 	opl17:
 		agetw();
-		while(tok.i >= 0) {
+		while((short) tok.i >= 0) {
 			outb(TYPEABS,tok.u & 0377);
 			agetw();
 		}
